refactor: ParentService 리팩토링
* @Transactional(readOnly = true) 어노테이션을 클래스 레벨로 변경
