+++ BE/src/main/java/ifive/idrop/driver/dto/DriverSubscribeInfoResponse.java
@@ -1,5 +1,6 @@
 package ifive.idrop.driver.dto;
 
+import ifive.idrop.child.domain.Child;
 import ifive.idrop.entity.*;
 import ifive.idrop.parent.domain.Parent;
 import lombok.Builder;
